requirements.txt
requests >= 2.3.0
curlify >= 2.1.0
playwright >= 1.49.0
python-dateutil >= 2.9.0.post0